# impala-bridge/openapi.yaml
openapi: 3.1.0
info:
  title: impala-bridge
  version: "0.0.0"
  summary: REST bridge between offline Payala payments and the Stellar network.
  description: |
    impala-bridge issues JWTs, persists dual-chain transactions, dispatches
    notifications via SMS/email/push/webhook, and streams events from Stellar
    Horizon and Payala.

    Authentication uses a two-token strategy:

    - **Refresh token** (14-day TTL) — obtained via username+password from `POST /token`.
      Used only to mint temporal tokens, never for API access.
    - **Temporal token** (1-hour TTL) — obtained from a refresh token via `POST /token`.
      Sent as `Authorization: Bearer <token>` on every authenticated request.

    All mutating endpoints enforce account ownership via `require_owner`. A
    `POST /logout` revokes the current token (the JTI is added to a Redis
    blacklist until it would have expired naturally).

    Rate limits apply to `/authenticate`, `/token`, `/mfa/verify`, `/transaction`,
    and `/card`: 10 requests per 60-second window, keyed by account id (and
    MFA type where applicable). Exceeding the limit returns HTTP 429 with a
    `Retry-After` header.

    Global request timeout is 30 s (HTTP 408 on exceed). Postgres has a
    per-connection `statement_timeout` of 60 s.
  contact:
    name: impala-bridge maintainers
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
...
  - url: https://{host}
...
security:
  - bearerAuth: []
paths:
...
  /authenticate:
    post:
      tags: [auth]
      summary: Register or verify a user with username+password.
      description: |
        If no credentials exist for the account, registers the user with the
        supplied password (Argon2 hashed). If credentials exist, verifies the
        password. Returns a generic "Invalid credentials" for both missing
        accounts and wrong passwords (account-enumeration defense). Rate
        limited per account id (10/60 s) and per client source (30/60 s,
        shared across all pre-auth endpoints); lockout after 5 wrong
        passwords per (account, client source) for 15 minutes — unknown
        accounts never count toward it.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthenticateRequest' }
      responses:
        '200':
          description: Authentication outcome (success may be false with 200).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthenticateResponse' }
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /token:
    post:
      tags: [auth]
      summary: Mint or refresh JWTs.
      description: |
        Two flows. If `refresh_token` is supplied, returns a new refresh +
        temporal pair and revokes the presented refresh token. Otherwise the
        request must include `username`+`password` and returns only a refresh
        token. The password flow is rate limited per username (10/60 s) and
        per client source (30/60 s, shared across all pre-auth endpoints;
        refresh rotation is exempt), with a lockout after 5 wrong passwords
        per (username, client source) for 15 minutes — unknown or federated
        usernames get the same reply but never count toward it.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TokenRequest' }
      responses:
        '200':
          description: Token pair (or single refresh token for password flow).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401':
          description: Invalid or revoked refresh token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /session/login:
    post:
      tags: [auth]
      summary: Establish a browser cookie session from local credentials.
      description: |
        Same credential check and rate-limit/lockout regime as `POST /token`
        (password flow): per username (10/60 s), per client source (30/60 s,
        shared across all pre-auth endpoints), lockout after 5 wrong passwords
        per (username, client source). On success sets an HttpOnly session
        cookie and returns a CSRF synchronizer token to send as `X-CSRF-Token`
        on later unsafe-method cookie requests.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionLoginRequest' }
      responses:
        '200':
          description: Session established (also sets a `Set-Cookie` header).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Temporal token (1-hour TTL). Obtain via `POST /token` with a valid
        refresh token. Send as `Authorization: Bearer <token>`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: impala_session
      description: |
        Browser session cookie (HttpOnly), established by `POST /session/login`
        or `POST /auth/okta` with `cookie_mode: true`. In production the cookie
        is `__Host-`-prefixed (`__Host-impala_session`). Unsafe-method requests
        additionally require the CSRF synchronizer token in the `X-CSRF-Token`
        header.
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or revoked token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: "unauthorized" }
            message: { type: string }
          required: [code, message]
      required: [error]
    AuthenticateRequest:
      type: object
      properties:
        account_id: { type: string }
        password: { type: string, minLength: 8 }
      required: [account_id, password]
    AuthenticateResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        action:
          type: string
          enum: ["", registered, authenticated]
    TokenRequest:
      type: object
      description: Supply either `refresh_token` OR `username`+`password`.
      properties:
        username: { type: string, nullable: true }
        password: { type: string, nullable: true }
        refresh_token: { type: string, nullable: true }
    TokenResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        refresh_token: { type: string, nullable: true }
        temporal_token: { type: string, nullable: true }
...
    SessionLoginRequest:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
      required: [username, password]
    SessionResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        account_id: { type: string }
        is_admin: { type: boolean }
        csrf_token:
          type: string
          description: CSRF synchronizer token for `X-CSRF-Token` on unsafe-method cookie requests.
...